seperate action functionality into ListOrganisation and MyOrganisation components
