+++ src/domain/Organisations/ListOrganisations/ListOrganisations.js
@@ -4,13 +4,15 @@ import Search from "../../../components/Search/Search";
 import UserContext from "../../../context/UserContext/UserContext";
 import AccessDenied from "../../Error/AccessDenied/AccessDenied";
 import OrganisationService from "../../../services/OrganisationService/OrganisationService";
-import { grey } from "../../../settings";
+import { grey, green, red } from "../../../settings";
 import styled from "styled-components";
 import { breakpoint } from "../../../utils/breakpoint/breakpoint";
 import UserService from "../../../services/UserService/UserService";
 import { ReactComponent as ApproveCircle } from "./icons/approve-circle.svg";
 import { ReactComponent as DeclineCircle } from "./icons/decline-circle.svg";
 import { ReactComponent as Trash } from "./icons/trash.svg";
+import { toast } from "react-toastify";
+import ConfirmModal from "../../../components/ConfirmModal/ConfirmModal";
 
 const StyledActionDiv = styled.div`
   display: flex;
@@ -31,7 +33,18 @@ const StyledActionDiv = styled.div`
 const ListOrganisations = ({ location }) => {
   const { roles } = useContext(UserContext)[0];
 
+  const [selectedOrganisation, setSelectedOrganisation] = useState({});
+
   const [search, setSearch] = useState(false);
+
+  const [approveIsLoading, setApproveIsLoading] = useState(false);
+  const [approveModalIsOpen, setApproveModalIsOpen] = useState(false);
+
+  const [declineIsLoading, setDeclineIsLoading] = useState(false);
+  const [declineModalIsOpen, setDeclineModalIsOpen] = useState(false);
+
+  const [removeIsLoading, setRemoveIsLoading] = useState(false);
+  const [removeModalIsOpen, setRemoveModalIsOpen] = useState(false);
 
   const [data, setData] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
@@ -93,35 +106,110 @@ const ListOrganisations = ({ location }) => {
     fetchData();
   }, [search, setData, setIsLoading]);
 
-  function doApprove() {
-    alert("approve organisation");
+  function toggleApproveModal() {
+    if (approveIsLoading) return;
+
+    setApproveModalIsOpen(!approveModalIsOpen);
   }
-  function doDecline() {
-    alert("decline organisation");
+
+  function toggleDeclineModal() {
+    if (declineIsLoading) return;
+
+    setDeclineModalIsOpen(!declineModalIsOpen);
   }
-  function doRemove() {
-    alert("remove organisation");
+
+  function toggleRemoveModal() {
+    if (removeIsLoading) return;
+
+    setRemoveModalIsOpen(!removeModalIsOpen);
+  }
+
+  async function doApprove(reviewerMessage) {
+    if (approveIsLoading) return;
+
+    setApproveIsLoading(true);
+
+    selectedOrganisation.status = "published";
+    selectedOrganisation.reviewedAt = new Date();
+    selectedOrganisation.reviewerMessage = reviewerMessage;
+
+    const organisation = await OrganisationService.updateOrganisation(
+      selectedOrganisation.id,
+      selectedOrganisation
+    );
+
+    setApproveIsLoading(false);
+
+    if (organisation) {
+      toast.success(`${selectedOrganisation.name} was approved.`);
+    } else {
+      toast.error(`Unable to approve organisation.`);
+    }
+
+    setApproveModalIsOpen(false);
+  }
+
+  async function doDecline(reviewerMessage) {
+    if (declineIsLoading) return;
+    setDeclineIsLoading(true);
+
+    selectedOrganisation.status = "rejected";
+    selectedOrganisation.reviewedAt = new Date();
+    selectedOrganisation.reviewerMessage = reviewerMessage;
+
+    const organisation = await OrganisationService.updateOrganisation(
+      selectedOrganisation.id,
+      selectedOrganisation
+    );
+
+    setDeclineIsLoading(false);
+
+    if (organisation) {
+      toast.success(`${selectedOrganisation.name} was declined.`);
+    } else {
+      toast.error(`Unable to decline organisation.`);
+    }
+
+    setDeclineModalIsOpen(false);
+  }
+
+  async function doRemove() {
+    if (removeIsLoading) return;
+
+    setRemoveIsLoading(true);
+
+    const organisationDeleted = await OrganisationService.deleteOrganisation(
+      selectedOrganisation.id
+    );
+
+    setRemoveIsLoading(false);
+
+    if (organisationDeleted) {
+      toast.success(`${selectedOrganisation.name} removed.`);
+    } else {
+      toast.error(`Unable to remove organisation.`);
+    }
+
+    setRemoveModalIsOpen(false);
   }
 
   const actions = [
     {
       title: "Approve",
-      onClick: doApprove,
+      onClick: toggleApproveModal,
       icon: ApproveCircle,
     },
     {
       title: "Decline",
-      onClick: doDecline,
+      onClick: toggleDeclineModal,
       icon: DeclineCircle,
     },
     {
       title: "Remove",
-      onClick: doRemove,
+      onClick: toggleRemoveModal,
       icon: Trash,
     },
   ];
-
-  console.log(actions);
 
   const accessPermission = roles.includes("viewer") || roles.includes("admin");
 
@@ -141,7 +229,49 @@ const ListOrganisations = ({ location }) => {
         organisationUser={organisationUser}
         isLoading={isLoading}
         search={search}
+        setSelectedOrganisation={setSelectedOrganisation}
         actions={actions}
+      />
+      <ConfirmModal
+        isOpen={approveModalIsOpen}
+        toggleModal={toggleApproveModal}
+        confirmButtonLabel={"Approve"}
+        confirmButtonColor={green[400]}
+        borderColor={green[300]}
+        onConfirm={doApprove}
+        includeReviewerMessage={true}
+        reviewerMessagePlaceholder={
+          "Thank you for listing your organisation..."
+        }
+        confirmTitle={"Approve organisation"}
+      />
+      <ConfirmModal
+        isOpen={declineModalIsOpen}
+        toggleModal={toggleDeclineModal}
+        confirmButtonLabel={"Decline"}
+        confirmButtonColor={red[400]}
+        borderColor={red[400]}
+        onConfirm={doDecline}
+        includeReviewerMessage={true}
+        reviewerMessagePlaceholder={
+          "We can not add you right now for the following reasons..."
+        }
+        confirmTitle={"Decline organisation"}
+      />
+      <ConfirmModal
+        isOpen={removeModalIsOpen}
+        toggleModal={toggleRemoveModal}
+        confirmMessage={
+          <>
+            Are you sure you want to remove{" "}
+            <strong>{selectedOrganisation.name}</strong>?
+          </>
+        }
+        confirmButtonLabel={"Remove"}
+        confirmButtonColor={red[400]}
+        borderColor={red[400]}
+        onConfirm={doRemove}
+        includeReviewerMessage={false}
       />
     </>
   ) : (
